requirements.txt
# Delta CFO Agent - Required Dependencies
# Consolidated requirements including invoice processing module

# ============================================
# Core Web Framework
# ============================================
Flask>=2.0.0
gunicorn>=20.1.0
Werkzeug>=2.0.0
Flask-Limiter>=3.5.0     # Rate limiting for API routes
Flask-WTF>=1.2.1         # CSRF protection and form handling
Flask-Babel>=4.0.0       # Internationalization (i18n) support

# ============================================
# Data Processing & Analysis
# ============================================
pandas>=1.3.0
openpyxl>=3.0.0
xlrd>=2.0.0

# ============================================
# AI Integration
# ============================================
anthropic>=0.8.0

# ============================================
# Authentication & Authorization
# ============================================
firebase-admin>=6.4.0    # Firebase Admin SDK for server-side auth
PyJWT>=2.8.0             # JWT token handling and verification

# ============================================
# HTTP & Network
# ============================================
requests>=2.25.0

# ============================================
# Date & Time Utilities
# ============================================
python-dateutil>=2.8.0

# ============================================
# Invoice Processing - PDF Processing
# ============================================
PyMuPDF>=1.23.0          # PDF rendering and conversion (fitz)
PyPDF2>=3.0.0            # PDF text extraction
pdfplumber>=0.7.0        # Advanced PDF data extraction
pdf2image>=1.16.0        # PDF to image conversion
reportlab>=3.6.0         # PDF generation for testing

# ============================================
# Invoice Processing - Image Processing
# ============================================
Pillow>=10.0.0           # Image manipulation and format support
opencv-python>=4.8.0     # Advanced image processing

# ============================================
# Invoice Processing - OCR
# ============================================
pytesseract>=0.3.10      # Tesseract OCR wrapper

# ============================================
# Invoice Processing - Email Automation
# ============================================
exchangelib>=4.9.0       # Exchange server integration
imapclient>=2.3.1        # IMAP email client

# ============================================
# Email Service & Notifications
# ============================================
sendgrid>=6.11.0         # SendGrid email service
email-validator>=2.1.0   # Email validation

# ============================================
# File Type Detection & Handling
# ============================================
python-magic>=0.4.27     # File type detection

# ============================================
# Compressed File Support
# ============================================
py7zr>=0.20.0            # 7z archive support

# ============================================
# Fuzzy Matching & String Similarity
# ============================================
rapidfuzz>=3.0.0         # Fast fuzzy string matching

# ============================================
# Database Support - Cloud SQL PostgreSQL
# ============================================
psycopg2-binary>=2.9.9   # PostgreSQL adapter for Cloud SQL
flask-cors>=4.0.0        # CORS support for web APIs

# ============================================
# Google Cloud Dependencies
# ============================================
google-cloud-storage>=2.10.0           # Google Cloud Storage for file uploads
google-cloud-secret-manager>=2.16.4    # Secret management
cloud-sql-python-connector>=1.4.0      # Cloud SQL connector

# ============================================
# Environment & Configuration
# ============================================
python-dotenv>=1.0.0     # Environment variable management